Add files via upload
All team members have contributed to the same tasks of the project. This is due to the small group size and that each member wants experience which each aspect of the project including documentation, research, and component searching. We are also waiting on the parts to begin implementation, once this occurs, then individual tasks may be assigned.